[REFACTORY] request_data name refactoring
